changed so that user who sign in with github can still change location
place favouriterecursion function in function
