+++ client/src/pages/Settings.js
@@ -16,6 +16,11 @@ function Settings() {
 
   /* This part below is to handle form request */
   const [state, dispatch] = useStoreContext();
+  
+  const [newGithub, setNewGithub] = useState(state.user.github.username || "");
+  const [newUsername, setNewUsername] = useState(state.user.localusername);
+  const [newLanguages, setNewLanguages] = useState(state.languages);
+  const [newLocation, setNewLocation] = useState(state.location);
 
   const githubRef = useRef();
   const locationRef = useRef();
