+++ controllers/userController.js
@@ -118,7 +118,7 @@ module.exports = {
   },
 
   findGithubUser: function(req,res) {
-    db.User.find({"github.id": parseInt(req.params.user)})
+    db.User.find({"github.id": (req.params.user)})
     .then(dbModel => res.json(dbModel))
     .catch(err => res.status(422).json(err));
   }
